==> Ileum_scripts/07j_Ileum_AllCells_PredictionMapping_Visualize.Rmd <==
---
title: "Reference-based Label Prediction and Mapping: Visualization of Results"
author: "Jayne Wiarda"
date: "9Apr2021"
output: 
  github_document: 
    keep_html: TRUE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(fig.fullwidth=TRUE)
```

## Visualizing prediction & mapping results

### Load required software packages

For visualization, we rely primarily on the packages Seurat (Stuart & Butler et al. 2019) and ggplots (Wickham 2016). Refer to sessionInfo() at the bottom of the page for the R and package versions used.

Required packages: scales (Wickham & Seidel 2020), writexl (Ooms 2020), readxl (Wickham & Bryan 2019), dplyr (Wickham et al. 2020)

```{r, message = FALSE}
library(Seurat)
library(ggplot2)
library(scales)
library(writexl)  
library(readxl)
library(dplyr)
```

### Import relevant data

Read in our Seurat object:

```{r, warning = FALSE, error = FALSE}
il <- readRDS("/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/Seurat/IleumAtlasAll.rds") 
DefaultAssay(il) <- "RNA"
Idents(il) <- il$seurat_clusters 
```

Read in our prediction and mapping results:

```{r, warning = FALSE, error = FALSE}
HsIleumMapping <- as.data.frame(read_excel('/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/HsIleum_Elmentaite_MasterMeta_PredictionMapping.xlsx')) # read in human ileum mapping & prediction scores
colnames(HsIleumMapping) <-paste("HsIleum", colnames(HsIleumMapping), sep="_") # add dataset-specific prefix to column names
#colnames(HsIleumMapping) # look at column names
rownames(HsIleumMapping) <- HsIleumMapping$HsIleum_CellBarcodes # change row names to cell barcodes
HsIleumMapping <- subset(HsIleumMapping, select = -c(HsIleum_CellBarcodes)) # drop cell barcode column
```

```{r, warning = FALSE, error = FALSE}
MmIleumMapping <- as.data.frame(read_excel('/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/MmIleum_Xu_MasterMeta_PredictionMapping.xlsx')) # read in murine ileum mapping & prediction scores
colnames(MmIleumMapping) <-paste("MmIleum", colnames(MmIleumMapping), sep="_") # add dataset-specific prefix to column names
#colnames(MmIleumMapping) # look at column names
rownames(MmIleumMapping) <- MmIleumMapping$MmIleum_CellBarcodes # change row names to cell barcodes
MmIleumMapping <- subset(MmIleumMapping, select = -c(MmIleum_CellBarcodes)) # drop cell barcode column
```

```{r, warning = FALSE, error = FALSE}
SsPBMCMapping <- as.data.frame(read_excel('/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/SsPBMC_HerreraWiarda_MasterMeta_PredictionMapping.xlsx')) # read in human ileum mapping & prediction scores
colnames(SsPBMCMapping) <-paste("SsPBMC", colnames(SsPBMCMapping), sep="_") # add dataset-specific prefix to column names
#colnames(SsPBMCMapping) # look at column names
rownames(SsPBMCMapping) <- SsPBMCMapping$SsPBMC_CellBarcodes # change row names to cell barcodes
SsPBMCMapping <- subset(SsPBMCMapping, select = -c(SsPBMC_CellBarcodes)) # drop cell barcode column
```

### Add mapping and prediction results as meta data:

First place cells in same order as listed in Seurat object:

```{r, warning = FALSE, error = FALSE}
HsIleumMapping <- HsIleumMapping[match(colnames(il), rownames(HsIleumMapping)),]
MmIleumMapping <- MmIleumMapping[match(colnames(il), rownames(MmIleumMapping)),]
SsPBMCMapping <- SsPBMCMapping[match(colnames(il), rownames(SsPBMCMapping)),]
```

Then add new data to the Seurat object: 

```{r, warning = FALSE, error = FALSE}
il <- AddMetaData(il, c(HsIleumMapping, MmIleumMapping, SsPBMCMapping))
```

### Overlay data onto 2-D visualizations

Visualize mapping scores overlaid onto UMAP:

```{r, warning = FALSE, error = FALSE}
FeaturePlot(il, 
            features = c('SsPBMC_MappingScores', 
                         'MmIleum_MappingScores', 
                         'HsIleum_MappingScores'),
            reduction = 'umap', # change to 'tsne' to overlay onto t-SNE plot
            ncol = 3) & 
  scale_color_gradientn( colours = c('yellow', 'orange', 'red', 'red4'),  limits = c(0.5, 1), oob = squish) & 
  NoAxes() & NoLegend()
```

Visualize lineage prediction scores overlaid onto UMAP:

```{r, warning = FALSE, error = FALSE}
FeaturePlot(il, 
            features = c('SsPBMC_CellLineage_prediction.score.T.ILC.lineage', 
                         'SsPBMC_CellLineage_prediction.score.B.lineage', 
                         'SsPBMC_CellLineage_prediction.score.Myeloid.lineage',
                         'SsPBMC_CellLineage_prediction.score.Non.immune',
                         'HsIleum_CellLineage_prediction.score.T.ILC.lineage', 
                         'HsIleum_CellLineage_prediction.score.B.lineage', 
                         'HsIleum_CellLineage_prediction.score.Myeloid.lineage',
                         'HsIleum_CellLineage_prediction.score.Non.immune',
                         'MmIleum_CellLineage_prediction.score.T.ILC.lineage', 
                         'MmIleum_CellLineage_prediction.score.B.lineage', 
                         'MmIleum_CellLineage_prediction.score.Myeloid.lineage',
                         'MmIleum_CellLineage_prediction.score.Non.immune'),
            reduction = 'umap', # change to 'tsne' to overlay onto t-SNE plot
            ncol = 4) & 
  scale_color_gradientn( colours = c('grey90', 'blue3'),  limits = c(0, 1), oob = squish) & 
  NoAxes() & NoLegend()
```

Visualize other prediction scores to mouse data overlaid onto UMAP:

```{r, warning = FALSE, error = FALSE, fig.width = 30}
MmPredict <- colnames(il@meta.data %>% select(starts_with("MmIleum_CellType_prediction.score."))) # extract 
FeaturePlot(il,
            features = c(MmPredict),
            reduction = 'umap', # change to 'tsne' to overlay onto t-SNE plot
            ncol = 8) & 
  scale_color_gradientn( colours = c('grey90', 'blue3'),  limits = c(0, 1), oob = squish) & 
  NoAxes() & NoLegend() 
```

I didn't take time to fix all the titles here, but left to right, top-to-bottom, these are the values being plotted:

```{r, warning = FALSE, error = FALSE}
MmPredict
```

Visualize other prediction scores to human data overlaid onto UMAP:

```{r, warning = FALSE, error = FALSE, fig.width = 30}
HsPredict <- colnames(il@meta.data %>% select(starts_with("HsIleum_CellType_prediction.score."))) # extract 
FeaturePlot(il,
            features = c(HsPredict),
            reduction = 'umap', # change to 'tsne' to overlay onto t-SNE plot
            ncol = 10) & 
  scale_color_gradientn( colours = c('grey90', 'blue3'),  limits = c(0, 1), oob = squish) & 
  NoAxes() & NoLegend()
```

I didn't take time to fix all the titles here, but left to right, top-to-bottom, these are the values being plotted:

```{r, warning = FALSE, error = FALSE}
HsPredict
```

Visualize other prediction scores to pig data overlaid onto UMAP:

```{r, warning = FALSE, error = FALSE, fig.width = 30}
SsPredict <- colnames(il@meta.data %>% select(starts_with("SsPBMC_CellType_prediction.score."))) # extract 
FeaturePlot(il,
            features = c(SsPredict),
            reduction = 'umap', # change to 'tsne' to overlay onto t-SNE plot
            ncol = 7) & 
  scale_color_gradientn( colours = c('grey90', 'blue3'),  limits = c(0, 1), oob = squish) & 
  NoAxes() & NoLegend() 
SsTitles <- sub(".*SsPBMC_CellType_prediction.score.", "", SsPredict)
```

I didn't take time to fix all the titles here, but left to right, top-to-bottom, these are the values being plotted:

```{r, warning = FALSE, error = FALSE}
SsPredict
```

There are a lots of different predictions going on, so let's try to summarize these into a single dot plot rather than looking at all of these feature plots...

### Integrate data into dot plots

#### Porcine PBMC mapping/prediction data

Let's make a dot plot summarizing porcine PBMC cell type predictions, summarized at the cluster level of our porcine ileum data.

Start by gathering all of the data required and calculating proportions of cells from query clusters assigned to particular reference cell types:

```{r, warning = FALSE, error = FALSE}
out <- data.frame("PigIleumCluster" = il$seurat_clusters, "PredictedPorcinePBMCCellType" = il$SsPBMC_CellType_predicted.id, "MappingScore" = il$SsPBMC_MappingScores) # make dataframe of cell cluster IDs, predicted cell types, and mapping scores
data <- table("PigIleumCluster" = il$seurat_clusters, "PredictedPorcinePBMCID" = il$SsPBMC_CellType_predicted.id) # create a table of # of cells from each cluster predicted as a certain cell type
data <- prop.table(data,1) # make #s in previous table into proportions
#rowSums(data) # should all equal 1
#head(data)
data <- as.data.frame(data) # convert table to data frame
head(data) # this is what the data looks like now
```

Now calculate average mapping scores for cells from each cluster assigned to a particular reference cell type: 

```{r, warning = FALSE, error = FALSE}
results <- list() 
for(i in 1:nrow(data)) {
  sub <- subset(out, PigIleumCluster == data[i,1])
  sub <- subset(sub, PredictedPorcinePBMCCellType == data[i,2])
  score <- as.data.frame(mean(sub$MappingScore))
  score$PigIleumCluster <- data[i,1]
  score$PredictedPorcinePBMCID <- data[i,2]
  score$Freq <- data[i,3]
  results[[i]] <- score
}

PredictOut <- do.call(rbind.data.frame, results)
#colnames(PredictOut)
names(PredictOut)[1] <- "MappingScore"
#colnames(PredictOut)
PredictOut[PredictOut=="NaN"]<-0 # convert all NaN values to quantitative zeros
PredictOut$Score <- as.numeric(PredictOut$Score)
PredictOut$Freq <- as.numeric(PredictOut$Freq)
PredictOut$PigIleumCluster <- as.character(PredictOut$PigIleumCluster)
```

Specify the order of variables to plot on our axes:

```{r, warning = FALSE, error = FALSE}
Idents(il) <- il$neworder
PredictOut$PigIleumCluster <- factor(PredictOut$PigIleumCluster, levels=c(levels(Idents(il)))) # order clusters in the sequence specified by 'neworder'
PredictOut$PredictedPorcinePBMCID <- factor(PredictOut$PredictedPorcinePBMCID, levels=c('B cells', 'Antibody-secreting cells', 'CD4pos ab T cells',
                                                                                      'CD8ABpos ab T cells', 'CD8Apos ab T/NK cells', 'NK cells',
                                                                                      'CD2pos gd T cells', 'CD2neg gd T cells', 'Monocytes', 'cDC', 
                                                                                      'pDC', 'Erythrocytes')) # specify order of reference cell types
```

Go ahead and plot the data:

```{r, warning = FALSE, error = FALSE, fig.width = 20}
ggplot(PredictOut, aes(x=PigIleumCluster, y=PredictedPorcinePBMCID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

There's definitely some 'noise' in this plot, so let's eliminate any dots that had very few cells from a cluster (<5% of cells from a cluster):

```{r, warning = FALSE, error = FALSE, fig.width = 20}
PredictOut2 <- PredictOut
PredictOut2$Freq[PredictOut2$Freq<.05] <- NA # if prediction was in less than 5% of cells from a cluster, set to NA so a dot no longer appears
ggplot(PredictOut2, aes(x=PigIleumCluster, y=PredictedPorcinePBMCID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

Much better!

Now let's also save the summarized data that we plotted:

```{r, warning = FALSE, error = FALSE}
write_xlsx(PredictOut, '/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/SsPBMC_HerreraWiarda_CellType_MappingResults.xlsx')
```

#### Human ileum mapping/prediction data

Let's make a dot plot summarizing human ileum cell type predictions, summarized at the cluster level of our porcine ileum data.

Start by gathering all of the data required and calculating proportions of cells from query clusters aHsigned to particular reference cell types:

```{r, warning = FALSE, error = FALSE}
out <- data.frame("PigIleumCluster" = il$seurat_clusters, "PredictedHumanIleumCellType" = il$HsIleum_CellType_predicted.id, "MappingScore" = il$HsIleum_MappingScores) # make dataframe of cell cluster IDs, predicted cell types, and mapping scores
data <- table("PigIleumCluster" = il$seurat_clusters, "PredictedHumanIleumID" = il$HsIleum_CellType_predicted.id) # create a table of # of cells from each cluster predicted as a certain cell type
data <- prop.table(data,1) # make #s in previous table into proportions
#rowSums(data) # should all equal 1
#head(data)
data <- as.data.frame(data) # convert table to data frame
head(data) # this is what the data looks like now
```

Now calculate average mapping scores for cells from each cluster aHsigned to a particular reference cell type: 

```{r, warning = FALSE, error = FALSE}
results <- list() 
for(i in 1:nrow(data)) {
  sub <- subset(out, PigIleumCluster == data[i,1])
  sub <- subset(sub, PredictedHumanIleumCellType == data[i,2])
  score <- mean(sub$MappingScore)
  score$PigIleumCluster <- data[i,1]
  score$PredictedHumanIleumID <- data[i,2]
  score$Freq <- data[i,3]
  results[[i]] <- score
}

PredictOut <- do.call(rbind.data.frame, results)
#colnames(PredictOut)
names(PredictOut)[1] <- "MappingScore"
#colnames(PredictOut)
PredictOut[PredictOut=="NaN"]<-0 # convert all NaN values to quantitative zeros
PredictOut$Score <- as.numeric(PredictOut$Score)
PredictOut$Freq <- as.numeric(PredictOut$Freq)
PredictOut$PigIleumCluster <- as.character(PredictOut$PigIleumCluster)
```

Specify the order of variables to plot on our axes:

```{r, warning = FALSE, error = FALSE}
Idents(il) <- il$neworder
PredictOut$PigIleumCluster <- factor(PredictOut$PigIleumCluster, levels=c(levels(Idents(il)))) # order clusters in the sequence specified by 'neworder'
PredictOut$PredictedHumanIleumID <- factor(PredictOut$PredictedHumanIleumID, levels=c("B cell", "Cycling B cell", "FCER2 B cell", "Memory B cell", 
                                                                                      "Activated B cell",  "IgA plasma cell", "IgG plasma cell",
                                                                                      "CD4 T cell", "Tfh", "Treg", "CD8 T cell", "gd T/NK cell", "Activated T", 
                                                                                      "Macrophage","Monocyte","cDC1", "cDC2","Cycling myeloid cells",  "mast cells",
                                                                                      "early enterocyte", "enterocyte", "BEST4 enterocyte", "crypt", "Goblet cell",
                                                                                      "Paneth cell", "enteroendocrine", "TA", "Arterial endothelial cell",
                                                                                      "Lymphatic endothelial cell","Venous endothelial cell","myofibroblast",  
                                                                                      "pericyte", "S1 fibroblasts", "S2 fibroblasts","Glial cell"))# specify order of reference cell types
```

Go ahead and plot the data:

```{r, warning = FALSE, error = FALSE, fig.width = 20}
ggplot(PredictOut, aes(x=PigIleumCluster, y=PredictedHumanIleumID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

There's definitely some 'noise' in this plot, so let's eliminate any dots that had very few cells from a cluster (<5% of cells from a cluster):

```{r, warning = FALSE, error = FALSE, fig.width = 20}
PredictOut2 <- PredictOut
PredictOut2$Freq[PredictOut2$Freq<.05] <- NA # if prediction was in leHs than 5% of cells from a cluster, set to NA so a dot no longer appears
ggplot(PredictOut2, aes(x=PigIleumCluster, y=PredictedHumanIleumID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

Much better!

Now let's also save the summarized data that we plotted:

```{r, warning = FALSE, error = FALSE}
write_xlsx(PredictOut, '/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/HsIleum_Elmentaite_CellType_MappingResults.xlsx')
```

#### Murine Ileum mapping/prediction data

Let's make a dot plot summarizing Murine Ileum cell type predictions, summarized at the cluster level of our Murine ileum data.

Start by gathering all of the data required and calculating proportions of cells from query clusters aMmigned to particular reference cell types:

```{r, warning = FALSE, error = FALSE}
out <- data.frame("PigIleumCluster" = il$seurat_clusters, "PredictedMurineIleumCellType" = il$MmIleum_CellType_predicted.id, "MappingScore" = il$MmIleum_MappingScores) # make dataframe of cell cluster IDs, predicted cell types, and mapping scores
data <- table("PigIleumCluster" = il$seurat_clusters, "PredictedMurineIleumID" = il$MmIleum_CellType_predicted.id) # create a table of # of cells from each cluster predicted as a certain cell type
data <- prop.table(data,1) # make #s in previous table into proportions
#rowSums(data) # should all equal 1
#head(data)
data <- as.data.frame(data) # convert table to data frame
head(data) # this is what the data looks like now
```

Now calculate average mapping scores for cells from each cluster aMmigned to a particular reference cell type: 

```{r, warning = FALSE, error = FALSE}
results <- list() 
for(i in 1:nrow(data)) {
  sub <- subset(out, PigIleumCluster == data[i,1])
  sub <- subset(sub, PredictedMurineIleumCellType == data[i,2])
  score <- mean(sub$MappingScore)
  score$PigIleumCluster <- data[i,1]
  score$PredictedMurineIleumID <- data[i,2]
  score$Freq <- data[i,3]
  results[[i]] <- score
}

PredictOut <- do.call(rbind.data.frame, results)
#colnames(PredictOut)
names(PredictOut)[1] <- "MappingScore"
#colnames(PredictOut)
PredictOut[PredictOut=="NaN"]<-0 # convert all NaN values to quantitative zeros
PredictOut$Score <- as.numeric(PredictOut$Score)
PredictOut$Freq <- as.numeric(PredictOut$Freq)
PredictOut$PigIleumCluster <- as.character(PredictOut$PigIleumCluster)
```

Specify the order of variables to plot on our axes:

```{r, warning = FALSE, error = FALSE}
Idents(il) <- il$neworder
PredictOut$PigIleumCluster <- factor(PredictOut$PigIleumCluster, levels=c(levels(Idents(il)))) # order clusters in the sequence specified by 'neworder'
PredictOut$PredictedMurineIleumID <- factor(PredictOut$PredictedMurineIleumID, levels=c("Resting B cell", "GC B cell (LZ)", "GC B cell (DZ)", "Plasma cell" ,
                                                                                        "Resting CD4pos T cell", "Activated CD4pos T cell", "CD8T T cell", 
                                                                                        "gd T cel (GZMApos)", "gd T cell (XCL1pos)", "NKT cell", "NK cell", 
                                                                                        "ILC1", "ILC2", "ILC3", "LTi cell", 
                                                                                        "Macrophage", "DC (CD103neg C1)", "DC (CD103neg C2)", "DC (CD103posCD11bneg)",
                                                                                        "DC (CD103posCD11bpos)", "pDC", "Mast cell", "Basophil", "Neutrophil",
                                                                                        "Epithelial cell C1", "Epithelial cell C2", "Endothelial cell", 
                                                                                        "Lymphatic endothelial-like cell", "Stromal cell (DN)", "Fibroblast")) # specify order of reference cell types
```

Go ahead and plot the data:

```{r, warning = FALSE, error = FALSE, fig.width = 20}
ggplot(PredictOut, aes(x=PigIleumCluster, y=PredictedMurineIleumID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

There's definitely some 'noise' in this plot, so let's eliminate any dots that had very few cells from a cluster (<5% of cells from a cluster):

```{r, warning = FALSE, error = FALSE, fig.width = 20}
PredictOut2 <- PredictOut
PredictOut2$Freq[PredictOut2$Freq<.05] <- NA # if prediction was in leMm than 5% of cells from a cluster, set to NA so a dot no longer appears
ggplot(PredictOut2, aes(x=PigIleumCluster, y=PredictedMurineIleumID, size=Freq, color=MappingScore)) + 
  geom_point() + 
  scale_color_gradientn(colours = c('yellow','orange','red', 'darkred'), limits = c(0.5, 1), oob = squish)+ 
  theme_bw()
```

Much better!

Now let's also save the summarized data that we plotted:

```{r, warning = FALSE, error = FALSE}
write_xlsx(PredictOut, '/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/MappingResults/MmIleum_Xu_CellType_MappingResults.xlsx')
```

### Save our data

Since we've added some new meta data information to our dataset, let's save our Seurat object as an .rds again:

```{r, warning = FALSE, error = FALSE, eval = TRUE}
saveRDS(il, "/home/Jayne.Wiarda/scRNAseqIleumAtlas/Ileum/Seurat/IleumAtlasAll.rds")
```
  
### View session information

```{r, warning = FALSE, error = FALSE}
sessionInfo()
```

### References

Jeroen Ooms (2020). writexl: Export Data Frames to Excel 'xlsx' Format. R package version 1.3.1. https://CRAN.R-project.org/package=writexl

Stuart and Butler et al. Comprehensive Integration of Single-Cell Data. Cell (2019).

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Hadley Wickham and Dana Seidel (2020). scales: Scale Functions for Visualization. R package version 1.1.1. https://CRAN.R-project.org/package=scales

Hadley Wickham and Jennifer Bryan (2019). readxl: Read Excel Files. R package version 1.3.1. https://CRAN.R-project.org/package=readxl

Hadley Wickham and Dana Seidel (2020). scales: Scale Functions for Visualization. R package version 1.1.1. https://CRAN.R-project.org/package=scales

Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020). dplyr: A Grammar of Data Manipulation. R package version 1.0.2. https://CRAN.R-project.org/package=dplyr
